test: move retention test to dedicated file
